fix : philo.h

+++ philo/philo_utils.c
@@ -6,7 +6,7 @@
 /*   By: yotak <yotak@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/26 15:32:48 by yotak             #+#    #+#             */
-/*   Updated: 2022/06/26 15:35:17 by yotak            ###   ########.fr       */
+/*   Updated: 2022/06/26 15:51:46 by yotak            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -126,7 +126,7 @@ int is_all_eat(t_info *info)
 	idx = 0;
 	while(idx < info->nbr_philos)
 	{
-		if (info->philo[idx].eat_cnt < info->nbr_eat_must)
+		if (info->philo[idx].eat_cnt < (int)info->nbr_eat_must)
 			return (FALSE);
 		idx += 1;
 	}
